Created a multi-targeted .NET libraries Supports both .Net Standard 2.0 and .Net Framework V4.5+
