+++ example/SeqLog4NetExample/App.config
@@ -4,7 +4,7 @@
     <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"/>
   </configSections>
   <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8"/>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
   </startup>
   <log4net>
     <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender">
