+++ example/SeqLog4NetExampleNetCore/SeqLog4NetExampleNetCore.csproj
@@ -1,12 +1,12 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Seq.Client.Log4Net.NetStandard\Seq.Client.Log4Net.NetStandard.csproj" />
+    <ProjectReference Include="..\..\src\Seq.Client.Log4Net\Seq.Client.Log4Net.csproj" />
   </ItemGroup>
 
   <ItemGroup>
