+++ example/SeqLog4NetExampleNetCore/log4net.config
@@ -2,8 +2,8 @@
 <log4net>
   <appender name="SeqAppender" type="Seq.Client.Log4Net.SeqAppender, Seq.Client.Log4Net" >
     <bufferSize value="1" />
-    <serverUrl value="https://" />
-    <apiKey value="LMmISXhwIlpf39Om" />
+    <serverUrl value="https://localhost:44391/" />
+    <apiKey value="6GK2tJ9hKZvw2uU7bR02" />
     <parameter>
       <parameterName value="Par1"/>
       <layout type="log4net.Layout.PatternLayout">
