+++ seq-client-log4net.sln
@@ -8,8 +8,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{5EB735
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Seq.Client.Log4Net", "src\Seq.Client.Log4Net\Seq.Client.Log4Net.csproj", "{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeqLog4NetExample", "example\SeqLog4NetExample\SeqLog4NetExample.csproj", "{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{50765ADD-1BED-452B-BC5A-4B1BFF527872}"
@@ -26,9 +24,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "sln", "sln", "{22E15F9D-1EE
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "example", "example", "{DC877E7E-1844-4670-8C26-77791A304EBD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Seq.Client.Log4Net.NetStandard", "src\Seq.Client.Log4Net.NetStandard\Seq.Client.Log4Net.NetStandard.csproj", "{734AD89C-968C-4D0A-A120-F922AD9DFAA1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SeqLog4NetExampleNetCore", "example\SeqLog4NetExampleNetCore\SeqLog4NetExampleNetCore.csproj", "{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeqLog4NetExampleNetCore", "example\SeqLog4NetExampleNetCore\SeqLog4NetExampleNetCore.csproj", "{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Seq.Client.Log4Net", "src\Seq.Client.Log4Net\Seq.Client.Log4Net.csproj", "{9445E368-7D79-4FB6-B515-BDDE5C73D41B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,31 +34,26 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16}.Release|Any CPU.Build.0 = Release|Any CPU
-		{734AD89C-968C-4D0A-A120-F922AD9DFAA1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{734AD89C-968C-4D0A-A120-F922AD9DFAA1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{734AD89C-968C-4D0A-A120-F922AD9DFAA1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{734AD89C-968C-4D0A-A120-F922AD9DFAA1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9445E368-7D79-4FB6-B515-BDDE5C73D41B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9445E368-7D79-4FB6-B515-BDDE5C73D41B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9445E368-7D79-4FB6-B515-BDDE5C73D41B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9445E368-7D79-4FB6-B515-BDDE5C73D41B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{CBDC0A30-5EC1-4DFE-AB0A-8CFC181B7DAB} = {50765ADD-1BED-452B-BC5A-4B1BFF527872}
 		{EE667DA8-C6CD-4238-8AB0-DC438A8DAB16} = {DC877E7E-1844-4670-8C26-77791A304EBD}
-		{734AD89C-968C-4D0A-A120-F922AD9DFAA1} = {50765ADD-1BED-452B-BC5A-4B1BFF527872}
 		{BA9EFB14-3AA5-4924-B785-51F2CC92A0D3} = {DC877E7E-1844-4670-8C26-77791A304EBD}
+		{9445E368-7D79-4FB6-B515-BDDE5C73D41B} = {50765ADD-1BED-452B-BC5A-4B1BFF527872}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3D1310A4-D8A9-4E11-A63B-A41947F53929}
