--- src/Seq.Client.Log4Net/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="log4net" version="2.0.8" targetFramework="net40" requireReinstallation="true" />
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net48" />
-  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net48" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="net48" />
-</packages>
